consolidate hook functions
