--- src/core/packages/notifications/browser-hooks/use_pinned_notifications.ts
@@ -1,18 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-import { useMemo } from 'react';
-import type { NotificationEvent } from '@kbn/core-notifications-browser';
-import { useNotifications } from './use_notifications';
-
-/** Notification events the user has pinned. */
-export function usePinnedNotifications(): NotificationEvent[] {
-  const all = useNotifications();
-  return useMemo(() => all.filter((e) => e.isPinned), [all]);
-}
